feat: html render for native (#2346)
