+++ apps/mobile/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@better-auth/expo": "1.0.23-beta.2",
     "@expo/vector-icons": "^14.0.2",
+    "@follow/components": "workspace:*",
     "@follow/constants": "workspace:*",
     "@follow/hooks": "workspace:*",
     "@follow/shared": "workspace:*",
@@ -83,6 +84,7 @@
     "drizzle-kit": "0.30.1",
     "eas-cli": "14.2.0",
     "expo-drizzle-studio-plugin": "0.1.1",
+    "postcss": "8.4.49",
     "react-test-renderer": "18.3.1"
   }
 }
