+++ apps/mobile/postcss.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  plugins: {
+    tailwindcss: {
+      config: "./tailwind.dom.config.ts",
+    },
+    autoprefixer: {},
+  },
+}
