+++ apps/mobile/src/components/common/Html.tsx
@@ -1,0 +1,14 @@
+"use dom"
+import "@follow/components/assets/colors-media.css"
+import "@follow/components/assets/tailwind.css"
+
+import type { HtmlProps } from "@follow/components"
+import { Html } from "@follow/components"
+
+export default function HtmlRender({
+  content,
+  dom,
+  ...options
+}: { dom?: import("expo/dom").DOMProps } & HtmlProps) {
+  return <Html content={content} {...options} />
+}
