+++ apps/renderer/package.json
@@ -103,12 +103,12 @@
     "sonner": "1.7.1",
     "tldts": "6.1.66",
     "unified": "11.0.5",
-    "unist-util-visit-parents": "^6.0.1",
+    "unist-util-visit-parents": "5.1.3",
     "use-context-selector": "2.0.0",
     "use-pull-to-refresh": "2.4.1",
     "use-sync-external-store": "1.4.0",
     "usehooks-ts": "3.1.0",
-    "vfile": "6.0.3",
+    "vfile": "5.3.7",
     "web-push": "3.6.7",
     "zod": "3.23.8",
     "zustand": "5.0.2"
