+++ package.json
@@ -140,13 +140,16 @@
       "jsonpointer": "patches/jsonpointer.patch",
       "workbox-precaching": "patches/workbox-precaching.patch",
       "@tanstack/react-virtual": "patches/@tanstack__react-virtual.patch",
-      "@pengx17/electron-forge-maker-appimage": "patches/@pengx17__electron-forge-maker-appimage.patch"
+      "@pengx17/electron-forge-maker-appimage": "patches/@pengx17__electron-forge-maker-appimage.patch",
+      "devlop": "patches/devlop.patch"
     },
     "overrides": {
       "@types/react": "npm:@types/react@^18.3.12",
       "@types/react-dom": "npm:@types/react-dom@^18.3.1",
       "is-core-module": "npm:@nolyfill/is-core-module@^1",
-      "isarray": "npm:@nolyfill/isarray@^1"
+      "isarray": "npm:@nolyfill/isarray@^1",
+      "unist-util-visit-parents": "5.1.3",
+      "vfile": "5.3.7"
     }
   },
   "simple-git-hooks": {
