+++ packages/components/assets/colors-media.css
@@ -1,0 +1,65 @@
+:root {
+  --fo-a: 21.6 100% 50%;
+
+  --fo-vibrancy-background: theme(colors.native.active);
+  --fo-vibrancy-foreground: 0 0% 45%;
+
+  --fo-text-primary: 0 0% 10%;
+
+  --fo-item-active: theme(colors.zinc.400/0.3);
+  --fo-item-hover: theme(colors.zinc.400/0.2);
+
+  --fo-inactive: 0 0% 80%;
+  --fo-disabled: 0 0% 70%;
+
+  --fo-background: theme(colors.background);
+
+  --fo-modal-background: theme(colors.zinc.50/0.8);
+  --fo-modal-background-opaque: theme(colors.zinc.50);
+
+  --fo-text-primary-hover: 0 0% 14.9%;
+
+  --fo-button-hover: theme(colors.zinc.500/0.1);
+  --fo-image-placeholder: theme(colors.zinc.100);
+  --fo-text-placeholder: theme(colors.zinc.400);
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --fo-a: 21.6 100% 50%;
+
+    --fo-vibrancy-foreground: 0 0% 83%;
+    --fo-vibrancy-background: theme(colors.native.active);
+
+    --fo-text-primary: 0 0% 80%;
+
+    --fo-item-active: theme(colors.neutral.600/0.4);
+    --fo-item-hover: theme(colors.neutral.700/0.3);
+
+    --fo-inactive: 0 0% 50%;
+    --fo-disabled: 0 0% 35%;
+
+    --fo-background: theme(colors.background);
+
+    --fo-modal-background: theme(colors.neutral.900/0.8);
+    --fo-modal-background-opaque: theme(colors.neutral.900);
+
+    --fo-text-primary-hover: 240 4.9% 83.9%;
+
+    --fo-button-hover: theme(colors.neutral.400/0.15);
+    --fo-image-placeholder: theme(colors.neutral.800);
+    --fo-text-placeholder: theme(colors.zinc.500);
+  }
+}
+
+:root {
+  --fo-native: 240 4.8% 95.9%;
+  --fo-native-active: 240 4.9% 83.9%;
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --fo-native: 220 8.1% 14.5%;
+    --fo-native-active: 198 31.3% 6.3%;
+  }
+}
