+++ packages/components/package.json
@@ -15,6 +15,7 @@
     "./ui/*": "./src/ui/*",
     "./utils/*": "./src/utils/*"
   },
+  "main": "./exports.ts",
   "peerDependencies": {
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
@@ -46,19 +47,31 @@
     "@radix-ui/react-toast": "1.2.2",
     "@radix-ui/react-tooltip": "1.1.4",
     "@react-hook/window-size": "3.1.1",
+    "@types/hast": "3.0.4",
+    "@types/unist": "3.0.3",
     "class-variance-authority": "0.7.1",
     "dayjs": "1.11.13",
     "foxact": "0.2.43",
     "framer-motion": "11.13.1",
+    "hast-util-to-jsx-runtime": "2.3.2",
+    "hast-util-to-text": "4.0.2",
     "jotai": "2.10.3",
     "masonic": "4.0.1",
     "react-blurhash": "^0.3.0",
     "react-fast-marquee": "1.6.5",
     "react-hook-form": "7.54.0",
     "react-i18next": "^15.1.3",
+    "rehype-infer-description-meta": "2.0.0",
+    "rehype-parse": "9.0.1",
+    "rehype-sanitize": "6.0.0",
+    "rehype-stringify": "10.0.1",
     "sonner": "1.7.1",
+    "unified": "11.0.5",
+    "unist-util-visit": "5.0.0",
+    "unist-util-visit-parents": "5.1.3",
     "use-context-selector": "2.0.0",
     "usehooks-ts": "3.1.0",
-    "vaul": "1.1.1"
+    "vaul": "1.1.1",
+    "vfile": "5.3.7"
   }
 }
