+++ patches/devlop.patch
@@ -1,0 +1,12 @@
+diff --git a/package.json b/package.json
+index 8319d8d58af116e5c7ae2b30ac214b05763b85f4..b788e890ece1329f24570f4e6376da669a2f813a 100644
+--- a/package.json
++++ b/package.json
+@@ -21,6 +21,7 @@
+   ],
+   "sideEffects": false,
+   "type": "module",
++  "main": "./lib/development.js",
+   "exports": {
+     "types": "./lib/development.d.ts",
+     "development": "./lib/development.js",
